Create utils file for enums and create_test_files script.
